+++ packages/atomic/src/components/search/result-lists/atomic-folded-result-list/atomic-folded-result-list.new.stories.tsx
@@ -29,7 +29,7 @@ export default meta;
 export const Default: Story = {
   name: 'atomic-folded-result-list',
   args: {
-    default: `
+    'slots-default': `
       <atomic-result-template>
         <template>
           <atomic-result-section-visual>