+++ packages/headless/src/features/commerce/common/actions.test.ts
@@ -7,6 +7,7 @@ import {buildMockCommerceFacetRequest} from '../../../test/mock-commerce-facet-r
 import {buildMockCommerceFacetSlice} from '../../../test/mock-commerce-facet-slice';
 import {buildMockCommerceRegularFacetValue} from '../../../test/mock-commerce-facet-value';
 import {buildMockCommerceState} from '../../../test/mock-commerce-state';
+import {VERSION} from '../../../utils/version';
 import {CommerceFacetSlice} from '../facets/facet-set/facet-set-state';
 import {
   getCommercePaginationInitialSlice,
@@ -53,6 +54,7 @@ describe('commerce common actions', () => {
               quantity: product.quantity,
             },
           ],
+          source: [`@coveo/headless@${VERSION}`, '@coveo/atomic@version'],
         },
       };
 
@@ -62,6 +64,9 @@ describe('commerce common actions', () => {
       state.configuration.accessToken = expected.accessToken;
       state.configuration.organizationId = expected.organizationId;
       state.configuration.analytics.trackingId = expected.trackingId;
+      state.configuration.analytics.source = {
+        '@coveo/atomic': 'version',
+      };
       state.commerceContext.language = expected.language;
       state.commerceContext.country = expected.country;
       state.commerceContext.currency = expected.currency as CurrencyCodeISO4217;